requirements.txt
jupyter_core
notebook>=5.6.0
python-Levenshtein>=0.12.2